api-normalizer/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-stdnum
python-multipart
scikit-learn
numpy